+++ frontend/src/pages/Dashboard.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { Box } from '@mui/material'
 import TopBar from 'components/Topbar'
-import WokerBottombar from './TransportWorker/Bottombar'
-import PickerBottombar from './Picker/Bottombar'
+import WokerBottombar from './TransportWorker/WokerBottomBar'
+import PickerBottombar from './Picker/PickerBottombar'
 import PendingTaskList from 'components/TaskList'
 import PickerCreatedTaskList from './Picker/TaskListCard'
 import { useAuth } from 'hooks/useAuth'
